+++ locksmith/src/controllers/v2/ticketsController.tsx
@@ -17,6 +17,7 @@ import {
 import { Verifier } from '../../models/verifier'
 import { getEventForLock } from '../../operations/eventOperations'
 import { notify } from '../../worker/helpers'
+import { isEventOrganizer } from '../../utils/getEventOrganizers'
 
 export class TicketsController {
   public web3Service: Web3Service
@@ -60,6 +61,7 @@ export class TicketsController {
       const verifier = eventSlug
         ? await Verifier.findOne({
             where: {
+              address,
               slug: eventSlug,
             },
           })
